requirements.txt
PyPDF2
azure-ai-formrecognizer==3.2.0b3
azure-common==1.1.28
azure-core==1.24.0
Flask
pdf2image
poppler-utils
pandas
xlrd==1.2.0